Updated user manual

+++ src/GameObject.h
@@ -18,7 +18,7 @@ public:
 	float width,height;//Dimensions of image
 	b2Body * body;//Pointer to box2d body
 	int depth;//For drawing order
-	int alpha;//Transparency
+	float alpha;//Transparency
 
 	static int numObjects;//Represents the number of objects created
